requirements.txt
aiogram
lxml
pymongo
python-docx
apscheduler
BeautifulSoup4
requests